TuVungAI/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
streamlit
requests
python-multipart
passlib[bcrypt]
python-dotenv
google-generativeai